+++ core/src/main/scala/hammerlab/iterator/package.scala
@@ -3,6 +3,7 @@ package hammerlab
 package object iterator
   extends bulk
      with count
+     with drop
      with scan
      with sliding
      with sorted
