+++ core/src/main/scala/org/hammerlab/iterator/scan/ScanValues.scala
@@ -1,10 +1,10 @@
 package org.hammerlab.iterator.scan
 
 import cats.Monoid
+import hammerlab.iterator.drop._
 import hammerlab.iterator.macros.IteratorWrapper
 import hammerlab.iterator.sliding._
 import org.hammerlab.iterator.{ MapValuesWithStateIterator, SimpleBufferedIterator }
-import org.hammerlab.iterator.DropRightIterator._
 
 @IteratorWrapper
 class ScanValues[K, V](it: Iterator[(K, V)]) {
